pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "A modular AI agent framework with ReAct pattern implementation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.32.5",
    "openai>=1.0.0",
    "anthropic>=0.25.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
]

[project.scripts]
ai-agent = "ai_agent.main:main"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.17.1",
    "pytest>=8.4.2",
    "ruff>=0.12.11",
    "jupyter>=1.0.0",
    "ipython>=8.0.0",
]
